Make it more efficient
This library was eating up all of my CPU and rendering pretty slowly. I cut out all of the cruft I didn't need and got everything running in a tighter loop. I'm not trying to save the world with this fork I'm just trying to give back the little that I can before I move on.